create strict typing for csv util
