+++ common/components/charts/SingleDayLineChart.tsx
@@ -63,7 +63,7 @@ export const SingleDayLineChart: React.FC<SingleDayLineProps> = ({
   pointField,
   benchmarkField,
   fname,
-  bothStops = false,
+  includeBothStopsForLocation = false,
   location,
   showLegend = true,
 }) => {
@@ -222,7 +222,7 @@ export const SingleDayLineChart: React.FC<SingleDayLineProps> = ({
               data={data}
               datasetName={fname}
               location={location}
-              bothStops={bothStops}
+              includeBothStopsForLocation={includeBothStopsForLocation}
               startDate={date}
             />
           )}
