+++ common/types/charts.ts
@@ -73,7 +73,7 @@ export interface LineProps {
   chartId: string;
   location: Location;
   pointField: PointField; // X value
-  bothStops?: boolean;
+  includeBothStopsForLocation?: boolean;
   fname: DataName;
   showLegend?: boolean;
 }
@@ -105,7 +105,7 @@ export interface HeadwayHistogramProps {
   date: string | undefined;
   location: Location;
   isLoading: boolean;
-  bothStops?: boolean;
+  includeBothStopsForLocation?: boolean;
   fname: DataName;
   showLegend?: boolean;
   metricField: MetricField;
