+++ modules/dwells/charts/DwellsAggregateChart.tsx
@@ -37,7 +37,7 @@ export const DwellsAggregateChart: React.FC<DwellsAggregateChartProps> = ({
         endDate={endDate}
         fillColor={CHART_COLORS.FILL}
         location={getLocationDetails(fromStation, toStation)}
-        bothStops={false}
+        includeBothStopsForLocation={false}
         fname="dwells"
       />
     );
