+++ modules/traveltimes/charts/TravelTimesAggregateChart.tsx
@@ -46,7 +46,7 @@ export const TravelTimesAggregateChart: React.FC<TravelTimesAggregateChartProps>
         endDate={endDate}
         fillColor={timeUnitByDate ? CHART_COLORS.FILL : CHART_COLORS.FILL_HOURLY}
         location={getLocationDetails(fromStation, toStation)}
-        bothStops={true}
+        includeBothStopsForLocation={true}
         fname="traveltimes"
       />
     );
